+++ package.json
@@ -17,7 +17,7 @@
     "format": "biome format --write src",
     "prepare": "husky",
     "test": "TZ=utc jest",
-    "generate:types": "npx openapi-typescript docs/api/openapi.yaml --output src/pkg/schema.gen.ts"
+    "generate:types": "npx openapi-typescript docs/api/openapi.yaml --output src/pkg/schema/schema.gen.ts"
   },
   "prisma": {
     "seed": "ts-node prisma/seed.ts"
